Handle writing a return
If there's no expression in the return we assume we will return 0.  It would probably be a better idea to construct a subroutine object and record the type at the start, so we could raise some kind of compilation error if it is a non-void subroutine and we've used `return;`.