Split parse_expression to handle terms and expressions separately
Specifically, when parsing a unaryOp we want the part after the unary operation to be a term, not an expression.  E.g. `-1 * 2` should `binaryOp(*, unaryOp(-, 1), 2)` not `unaryOp(-, binaryOp(*, 1, 2)).  If you actually want the second form you have to express it with parenthesis: `-(1 * 2)`.  The problem would manifest with the tokenizer being advanced too many times in the 2nd half of parse_expression and compilation of the next part of the code breaking.