Implement VM code generation for while statements
Again we drive this out from the VM code generated by the provided tools.  The labels might not be our preferred choice, but at least we have parity with the generated code for comparison.  The `build_label` / `reset_labels` / `all_labels` implementation probably wants to live in a separate object.  In fact it might make sense to extract the entire "compile_subroutine" process to another object, but I guess we have to use the API given us by the book.