Extract JackClass and JackMethod structs
We build these as we consume information from the source rather than passing variables around the methods.  The driver for this is that when we see a constructor we need to emit some more commands at the start of the VM code for the function definition to allocate memory.  I think we need to allocate memory for each field (don't have the book to hand so I'm just reverse engineering what the provided compiler does).