Handle method calls on "this"
If the method call is `<blah>(<params>?)` rather than `<class>.<blah>(<params>?)` then we have an implicit recipient of "this".  We still need to write the funcation call as `<class>.<blah>` though which means all our `#emit` calls need to be provided the class.  Happily we extracted that earlier and can add it to the arguments.  If we do have the "this" call we need to push pointer 0 onto the top of the stack as the first argument, much like we do with calling methods on variables.