Handle array references in let statements
We don't parse the variable reference using the ExpressionParser so we need to do some manual work to extract the array or standard variable reference.  We add .emit_assignment_to methods to Variable and ArrayReference because although assigning to a variable is just:

    VM code for expression
    pop variable

It's more complex for array assignment because we interleave the array reference and the expression we're assigning to it.  e.g.:

    get memory location (variable + index)
    flip pointer to that
    VM code for expression
    pop that

By adding this method we have the same API regardless of the type of variable we're assigning.

Not sure that we shouldn't push some of this work into another object, or the (increasingly poorly named) ExpressionParser as it feels a bit weird to construct a Variable or ArrayReference object outside of there.