Swap array variable and index expression when emitting index
In the book the example on page 229 shows "final VM code" as follows:

    // bar[k] = 19
    push local 2
    push argument 0
    add
    [snip: code for assignment skipped]

And makes reference to section 7.2.6 which on page 138 has code as follows:

    // bar[2] = 19
    push local 0
    push constant 2
    add
    [snip: code for assignment skipped]

So, naturally we assume that we output the address of the variable, the the index expression and then add them.  However, the provided VM tool does these in the opposite order; it emits the index expression, then the address of the variable then adds them.  It's not clear why this is in the opposite order, but in order to pass our specs where we compare our output with theirs we just have to follow along.

Idle thought: is it more "reverse polish" stuff? or perhaps to isolate the variable reference from anything that the index expression might do that would change what the variable address actually is (e.g. call a function that changes the variable somehow?)