Swap dereference and assignment expression for array assignment
Like last commit the book says one thing, but the VM code outputted by the tool says another.

In the book, on page 229 we have array reference assignment code as follows:

    // bar[k] = 19
    [snip: code for array reference skipped]
    // use the that segment to access bar[k]
    pop pointer 1
    push constant 19
    pop that 0

This again makes reference to section 7.2.6 which has similar code on page 138:

    // bar[2]=19
    [snip: code for array reference skipped]
    pop pointer 1 //set that's base to (bar+2)
    push constant 19
    pop that 0

In essence this tells us that our array assignment algorithm is:

1. calculate the array index memory location,
2. flip pointer to 1 so "that" is pointing at the calculated memory location
3. output the VM code for the value we want to assign to the array
4. write the value to "that"

However, the outputted tools actually do the following:

   // bar[2] = 19
   [snip: code for array reference skipped]
   push constant 19
   pop temp 0
   pop pointer 1
   push temp 0
   pop that 0

e.g the algorithm for assignment is:

1. calculate the array index memory location,
2. output the VM code for the value we want to assign to the array
3. store that result in temp
4. flip pointer to 1 so "that" is pointing at the calculated memory location
5. get the assignment value back from temp
6. write the value to "that"

Again, it's not clear *why* the provided tool does this, but we have to do the same to make our tests pass.

Idle thought: it seems like this might be about isolation, if we flip pointer to 1 before we run the assignment value then maybe that value would also change the value that pointer 1 points at.