Rework parsing of array references
An array reference is not an expression, it's a term.  This means reworking parse_expression and parse_term so that we look for [<expression>] inside parse_term, not inside parse_expression.

This means that we have to be more careful about when we advance the tokenizer and do it closer to when we've used the token.  E.g. the individual case statements in parse_term are responsible for advancing, we don't just always advance at the end.  We also make sure that parse_nested will advance after the final ')'.  parse_unary gets the correct behaviour because it falls through to parse_term.